refactor: 플레이어의 numberOfMyCards, getState, canDraw 함수 제거
