+++ src/main/kotlin/blackjack/service/FightService.kt
@@ -4,10 +4,11 @@ import blackjack.domain.Dealer
 import blackjack.domain.FightResult
 import blackjack.domain.Player
 import blackjack.domain.PlayerState
+import blackjack.error.BlackjackErrorMessage.DEALER_STATUS_CAN_NOT_BE_STAND
 
 class FightService {
     fun go(dealer: Dealer, player: Player): FightResult {
-        return when (dealer.getState()) {
+        return when (dealer.state) {
             PlayerState.BLACK_JACK -> dealerBlackjack(player)
             PlayerState.BUST -> dealerBust(dealer, player)
             PlayerState.HIT -> dealerHit(dealer, player)
@@ -16,7 +17,7 @@ class FightService {
     }
 
     private fun dealerBlackjack(player: Player): FightResult {
-        if (player.getState() == PlayerState.BLACK_JACK) {
+        if (player.state == PlayerState.BLACK_JACK) {
             return FightResult.DRAW
         }
 
@@ -24,7 +25,7 @@ class FightService {
     }
 
     private fun dealerBust(dealer: Dealer, player: Player): FightResult {
-        if (player.getState() == PlayerState.BUST) {
+        if (player.state == PlayerState.BUST) {
             return FightResult.LOSE
         }
 
@@ -36,7 +37,7 @@ class FightService {
     }
 
     private fun dealerHit(dealer: Dealer, player: Player): FightResult {
-        if (player.getState() == PlayerState.BUST) {
+        if (player.state == PlayerState.BUST) {
             return FightResult.LOSE
         }
