+++ src/test/kotlin/blackjack/PlayerTest.kt
@@ -21,20 +21,21 @@ class PlayerTest {
     @Test
     fun `플레이어는 카드들을 가질 수 있다`() {
         player1.hit()
-        assertThat(player1.numberOfMyCards()).isEqualTo(1)
+        val cards = player1.getMyCards()
+        assertThat(cards.cards.size).isEqualTo(1)
     }
 
     @Test
     fun `플레이어의 상태를 확인할 수 있다`() {
         player1.hit()
-        assertThat(player1.getState()).isEqualTo(PlayerState.HIT)
+        assertThat(player1.state).isEqualTo(PlayerState.HIT)
     }
 
     @Test
     fun `플레이어가 스탠드를 선언하면 더이상 카드를 뽑을 수 없다`() {
         player1.hit()
         player1.stand()
-        assertThat(player1.canDraw()).isEqualTo(false)
+        assertThat(player1.state.canDraw).isEqualTo(false)
     }
 
     @Test
